chore: RFCCard visual regressions and fixing tests
